STARK: Split detection code & adapt to new plugins.
