+++ engines/stark/detection.cpp
@@ -21,12 +21,7 @@
  */
 
 #include "engines/advancedDetector.h"
-#include "engines/stark/savemetadata.h"
-#include "engines/stark/stark.h"
-#include "engines/stark/services/stateprovider.h"
 
-#include "common/savefile.h"
-#include "common/system.h"
 #include "common/translation.h"
 
 namespace Stark {
@@ -370,9 +365,9 @@ static const ADExtraGuiOptionsMap optionsList[] = {
 	AD_EXTRA_GUI_OPTIONS_TERMINATOR
 };
 
-class StarkMetaEngine : public AdvancedMetaEngine {
+class StarkMetaEngineStatic : public AdvancedMetaEngineStatic {
 public:
-	StarkMetaEngine() : AdvancedMetaEngine(gameDescriptions, sizeof(ADGameDescription), starkGames, optionsList) {
+	StarkMetaEngineStatic() : AdvancedMetaEngineStatic(gameDescriptions, sizeof(ADGameDescription), starkGames, optionsList) {
 		_guiOptions = GUIO4(GUIO_NOMIDI, GAMEOPTION_ASSETS_MOD, GAMEOPTION_LINEAR_FILTERING, GAMEOPTION_FONT_ANTIALIASING);
 	}
 
@@ -387,97 +382,8 @@ public:
 	const char *getOriginalCopyright() const override {
 		return "(C) Funcom";
 	}
-
-	bool hasFeature(MetaEngineFeature f) const override {
-		return
-			(f == kSupportsListSaves) ||
-			(f == kSupportsLoadingDuringStartup) ||
-			(f == kSupportsDeleteSave) ||
-			(f == kSavesSupportThumbnail) ||
-			(f == kSavesSupportMetaInfo) ||
-			(f == kSavesSupportPlayTime) ||
-			(f == kSavesSupportCreationDate);
-	}
-
-	int getMaximumSaveSlot() const override {
-		return 999;
-	}
-
-	SaveStateList listSaves(const char *target) const override {
-		Common::StringArray filenames = StarkEngine::listSaveNames(target);
-
-		SaveStateList saveList;
-		for (Common::StringArray::const_iterator filename = filenames.begin(); filename != filenames.end(); ++filename) {
-			int slot = StarkEngine::getSaveNameSlot(target, *filename);
-
-			// Read the description from the save
-			Common::String description;
-			Common::InSaveFile *save = g_system->getSavefileManager()->openForLoading(*filename);
-			if (save) {
-				StateReadStream stream(save);
-				description = stream.readString();
-			}
-
-			saveList.push_back(SaveStateDescriptor(slot, description));
-		}
-
-		Common::sort(saveList.begin(), saveList.end(), SaveStateDescriptorSlotComparator());
-		return saveList;
-	}
-
-	SaveStateDescriptor querySaveMetaInfos(const char *target, int slot) const override {
-		Common::String filename = StarkEngine::formatSaveName(target, slot);
-		Common::InSaveFile *save = g_system->getSavefileManager()->openForLoading(filename);
-		if (!save) {
-			return SaveStateDescriptor();
-		}
-
-		SaveStateDescriptor descriptor;
-		descriptor.setSaveSlot(slot);
-
-		SaveMetadata metadata;
-		Common::ErrorCode readError = metadata.read(save, filename);
-		if (readError != Common::kNoError) {
-			delete save;
-			return descriptor;
-		}
-
-		descriptor.setDescription(metadata.description);
-
-		if (metadata.version >= 9) {
-			Graphics::Surface *thumb = metadata.readGameScreenThumbnail(save);
-			descriptor.setThumbnail(thumb);
-			descriptor.setPlayTime(metadata.totalPlayTime);
-			descriptor.setSaveDate(metadata.saveYear, metadata.saveMonth, metadata.saveDay);
-			descriptor.setSaveTime(metadata.saveHour, metadata.saveMinute);
-		}
-
-		if (metadata.version >= 13) {
-			descriptor.setAutosave(metadata.isAutoSave);
-		}
-
-		delete save;
-
-		return descriptor;
-	}
-
-	void removeSaveState(const char *target, int slot) const override {
-		Common::String filename = StarkEngine::formatSaveName(target, slot);
-		g_system->getSavefileManager()->removeSavefile(filename);
-	}
-
-	bool createInstance(OSystem *syst, Engine **engine, const ADGameDescription *desc) const override {
-		if (desc)
-			*engine = new StarkEngine(syst, desc);
-
-		return desc != nullptr;
-	}
 };
 
 } // End of namespace Stark
 
-#if PLUGIN_ENABLED_DYNAMIC(STARK)
-	REGISTER_PLUGIN_DYNAMIC(STARK, PLUGIN_TYPE_ENGINE, Stark::StarkMetaEngine);
-#else
-	REGISTER_PLUGIN_STATIC(STARK, PLUGIN_TYPE_ENGINE, Stark::StarkMetaEngine);
-#endif
+REGISTER_PLUGIN_STATIC(STARK_DETECTION, PLUGIN_TYPE_METAENGINE, Stark::StarkMetaEngineStatic);
